custard: regenerate against F* master, now that Custard is merged
FStarLang/FStar#4395 merged as FStarLang/FStar@d57a6d9ae3, so Custard is in
master and no longer needs a branch build.  Rebase onto fstar2 and regenerate
every snapshot against master.

The 37 commits since 09acba059a include Custard sections 115-118, which show
up in the output as four changes, all improvements:

  - The custard_unit typedef and the 16-bit float block are now emitted only
    into files that mention them, which neither COSE nor cbor does.  That is
    the entire COSE_Format.h diff: nine lines off the top of the header.
  - Shift operands are explicitly promoted -- `(uint8_t)((unsigned int)(x) <<
    (unsigned int)(8 - hi))` -- rather than relying on the implicit
    conversion.
  - Allocations carry an overflow check, `if ((sz) > SIZE_MAX / sizeof(T))
    abort();`.
  - `&&` now short-circuits.  Previously COSE_Format.c evaluated both
    operands of `(len == 0) && _ct1` before branching; it now computes the
    second only when the first holds.

No EverParse source needed a change: the C consumers, the cddl test drivers
and their #ifdef EVERPARSE_CUSTARD blocks all build untouched.

opt/hashes.Makefile is deliberately left alone.  CI builds the committed
snapshots against the pinned F*, which predates the merge, so the pin is the
automated advance job's to move; custard-detect.Makefile continues to pick
the backend from whichever compiler is at hand.

Verified green on both backends: COSE (C + Rust), cbor (363/363 det and
nondet in C, 724 via the Python cross-validation, 363+363 in Rust) and all
five cddl/tests corpora, with the karamel leg re-run as a control.

+++ src/cbor/pulse/nondet/c/CBORNondet.c
@@ -167,7 +167,7 @@ static uint8_t Pulse_Lib_Slice_op_Dot_Lparen_Rparen__t(Pulse_Lib_Slice_slice__ui
 }
 
 static uint8_t LowParse_BitFields_get_bitfield_gen8(uint8_t x, uint32_t lo, uint32_t hi) {
-  uint8_t op1 = ((uint8_t)(x << (8 - hi)));
+  uint8_t op1 = ((uint8_t)((unsigned int)(x) << (unsigned int)(8 - hi)));
   return (op1 >> ((8 - hi) + lo));
 }
 
@@ -1993,10 +1993,10 @@ size_t cbor_nondet_size(cbor_raw x, size_t bound) {
 static uint8_t LowParse_BitFields_set_bitfield_gen8(uint8_t x, uint32_t lo, uint32_t hi, uint8_t v) {
   uint8_t op0 = 255;
   uint8_t op1 = (op0 >> (8 - (hi - lo)));
-  uint8_t op2 = ((uint8_t)(op1 << lo));
+  uint8_t op2 = ((uint8_t)((unsigned int)(op1) << (unsigned int)(lo)));
   uint8_t op3 = ((uint8_t)(~(op2)));
   uint8_t op4 = (x & op3);
-  uint8_t op5 = ((uint8_t)(v << lo));
+  uint8_t op5 = ((uint8_t)((unsigned int)(v) << (unsigned int)(lo)));
   return (op4 | op5);
 }
